Correcting variable types
Earlier unpublished versions of the code used long int's instead of
double's as lc's general type for numbers. Therefore, the parser was
based on strtol() and function pointers in op.c/op.h were typed for long
int's. All instances of long int's have been corrected in this commit to
the intended double floating-point type intended.

+++ op.c
@@ -4,15 +4,15 @@
 
 #include "op.h"
 
-static long op_add(long m, long n);
+static double op_add(double m, double n);
 
 const Op op_list[] = {
 	{ "+", OP_N2, { .op_n2 = op_add } },
 	{ "", OP_N0, { .op_n0 = NULL } }
 };
 
-static long 
-op_add(long m, long n)
+static double
+op_add(double m, double n)
 {
 	return m + n;
 }
